Refactored: Moved `Core::Quantization` to `Contrib::Utils`
