+++ lib/ddtrace.rb
@@ -9,8 +9,6 @@ require 'ddtrace/version'
 require 'datadog/tracing'
 require 'ddtrace/pin'
 require 'ddtrace/tracer'
-require 'datadog/core/quantization/hash'
-require 'datadog/core/quantization/http'
 require 'ddtrace/pipeline'
 require 'ddtrace/configuration'
 require 'ddtrace/metrics'
@@ -20,6 +18,8 @@ require 'ddtrace/profiling'
 require 'datadog/contrib'
 require 'ddtrace/contrib/auto_instrument'
 require 'ddtrace/contrib/extensions'
+require 'ddtrace/contrib/utils/quantization/hash'
+require 'ddtrace/contrib/utils/quantization/http'
 
 require 'ddtrace/opentelemetry/extensions'
