+++ lib/ddtrace/contrib/mongodb/parsers.rb
@@ -1,4 +1,4 @@
-require 'datadog/core/quantization/hash'
+require 'ddtrace/contrib/utils/quantization/hash'
 
 # typed: true
 module Datadog
@@ -18,11 +18,11 @@ module Datadog
       # returns a formatted and normalized query
       def query_builder(command_name, database_name, command)
         # always exclude the command name
-        options = Core::Quantization::Hash.merge_options(quantization_options, exclude: [command_name.to_s])
+        options = Contrib::Utils::Quantization::Hash.merge_options(quantization_options, exclude: [command_name.to_s])
 
         # quantized statements keys are strings to avoid leaking Symbols in older Rubies
         # as Symbols are not GC'ed in Rubies prior to 2.2
-        base_info = Core::Quantization::Hash.format(
+        base_info = Contrib::Utils::Quantization::Hash.format(
           {
             'operation' => command_name,
             'database' => database_name,
@@ -31,11 +31,11 @@ module Datadog
           options
         )
 
-        base_info.merge(Core::Quantization::Hash.format(command, options))
+        base_info.merge(Contrib::Utils::Quantization::Hash.format(command, options))
       end
 
       def quantization_options
-        Core::Quantization::Hash.merge_options(DEFAULT_OPTIONS, configuration[:quantize])
+        Contrib::Utils::Quantization::Hash.merge_options(DEFAULT_OPTIONS, configuration[:quantize])
       end
 
       def configuration
