+++ lib/ddtrace/contrib/rake/instrumentation.rb
@@ -1,5 +1,5 @@
 # typed: false
-require 'datadog/core/quantization/hash'
+require 'ddtrace/contrib/utils/quantization/hash'
 require 'ddtrace/contrib/analytics'
 require 'ddtrace/contrib/rake/ext'
 
@@ -73,7 +73,7 @@ module Datadog
 
           def quantize_args(args)
             quantize_options = configuration[:quantize][:args]
-            Core::Quantization::Hash.format(args, quantize_options)
+            Contrib::Utils::Quantization::Hash.format(args, quantize_options)
           end
 
           def enabled?
