+++ spec/ddtrace/contrib/utils/quantization/http_spec.rb
@@ -1,9 +1,9 @@
 # typed: false
 require 'spec_helper'
 
-require 'datadog/core/quantization/http'
+require 'ddtrace/contrib/utils/quantization/http'
 
-RSpec.describe Datadog::Core::Quantization::HTTP do
+RSpec.describe Datadog::Contrib::Utils::Quantization::HTTP do
   describe '#url' do
     subject(:result) { described_class.url(url, options) }
